+++ pkg/analysis_server/lib/src/services/search/search_engine_internal.dart
@@ -356,6 +356,12 @@ class SearchMatchImpl implements SearchMatch {
     if (kind == SearchResultKind.INVOCATION) {
       return MatchKind.INVOCATION;
     }
+    if (kind == SearchResultKind.DOT_SHORTHANDS_CONSTRUCTOR_TEAR_OFF) {
+      return MatchKind.DOT_SHORTHANDS_CONSTRUCTOR_TEAR_OFF;
+    }
+    if (kind == SearchResultKind.DOT_SHORTHANDS_CONSTRUCTOR_INVOCATION) {
+      return MatchKind.DOT_SHORTHANDS_CONSTRUCTOR_INVOCATION;
+    }
     if (kind ==
         SearchResultKind.INVOCATION_BY_ENUM_CONSTANT_WITHOUT_ARGUMENTS) {
       return MatchKind.INVOCATION_BY_ENUM_CONSTANT_WITHOUT_ARGUMENTS;