+++ pkg/analysis_server/test/services/completion/dart/location/argument_list_test.dart
@@ -237,6 +237,40 @@ suggestions
     kind: keyword
 ''');
   }
+
+  Future<void> test_dotShorthands_constructor() async {
+    await computeSuggestions('''
+class A {
+  A({int? foo01, int? foo02});
+}
+
+A a = .new(^);
+''');
+    assertResponse(r'''
+suggestions
+  |foo01: |
+    kind: namedArgument
+  |foo02: |
+    kind: namedArgument
+''');
+  }
+
+  Future<void> test_dotShorthands_method() async {
+    await computeSuggestions('''
+class A {
+  static A foo({int? foo01, int? foo02}) => A();
+}
+
+A a = .foo(^);
+''');
+    assertResponse(r'''
+suggestions
+  |foo01: |
+    kind: namedArgument
+  |foo02: |
+    kind: namedArgument
+''');
+  }
 }
 
 @reflectiveTest