+++ pkg/analyzer_plugin/lib/src/utilities/completion/optype.dart
@@ -1837,6 +1837,10 @@ class _OpTypeAstVisitor extends GeneralizingAstVisitor<void> {
       var parent = node.parent;
       if (parent is Annotation) {
         return 'annotation';
+      } else if (parent is DotShorthandConstructorInvocation) {
+        return 'constructor';
+      } else if (parent is DotShorthandInvocation) {
+        return 'method';
       } else if (parent is EnumConstantArguments) {
         return 'enumConstantArguments';
       } else if (parent is ExtensionOverride) {